Use Livewire to add a regenerate button
