+++ resources/views/generate.blade.php
@@ -13,7 +13,7 @@
         <div class="text-center font-display text-white w-full font-display text-lg relative z-10">
             Synthwave Band Name Generator
         </div>
-        <h1 class="text-2xl md:text-3xl xl:text-4xl font-display text-white leading-tight text-shadow relative z-10 font-display py-8">{{ $name }}</h1>
+        @livewire('name')
         <a href="https://jackmcdade.com" class="font-display text-lg hover:text-purple text-pink relative z-10">Created by Jack McDade</a>
     </div>
     <script async src="https://www.googletagmanager.com/gtag/js?id=UA-145403615-1"></script>
@@ -23,5 +23,6 @@
     gtag('js', new Date());
     gtag('config', 'UA-145403615-1');
     </script>
+    @livewireAssets
 </body>
 </html>
