+++ resources/views/livewire/name.blade.php
@@ -1,0 +1,7 @@
+<div class="z-10">
+    <h1 class="text-2xl md:text-3xl xl:text-4xl font-display text-white leading-tight text-shadow relative z-10 font-display py-8">{{ $name }}</h1>
+
+    <div class="mb-4 text-center">
+        <button wire:click="generate">Another</button>
+    </div>
+</div>
